background → serviceWorker
only the badge is updated in the background,
we loose both search cache & icon cache update in the background (no access to fileSystem)
